DAPA-48 Update environment handling and remove unused main.py
Refactored `.env` file paths to use `app/secrets/.env` across the application. Removed `main.py` to align with the updated architecture. Simplified `vault-fetcher` logic to skip existing secrets and enhanced debug logging for better observability. Adjusted `docker-compose.yaml` and `.gitignore` accordingly.
